+++ src/views/admin/adminHome.php
@@ -3,7 +3,7 @@
 <div class="column is-12">
 
   <!-- Header -->
-  <h3 class="title has-text-black">[S11] Selection BTS - Espace Administrateur</h3>
+  <h3 class="title has-text-black">[S11] Selection BTS - Espace Administrateur - Accueil</h3>
   <hr class="login-hr">
   <p class="subtitle has-text-black">
     Vous pouvez consulter la liste des comptes et
@@ -28,6 +28,10 @@
         <a class="field button is-block is-dark is-medium" href="?logout">
           Se déconnecter
           <i class="mdi mdi-logout" aria-hidden="true"></i>
+        </a>
+        <a class="field button is-block is-dark is-medium" href="?">
+          Accueil
+          <i class="mdi mdi-home" aria-hidden="true"></i>
         </a>
         <a class="field button is-block is-dark is-medium" href="?create">
           Nouveau compte
@@ -76,6 +80,10 @@
 
   </div>
   <!-- ./Main content -->
+
+  <!-- Footer -->
+  <p class="has-text-grey">NOIZET Maxence &nbsp;·&nbsp; 2021</p>
+  <!-- ./Footer -->
 
 </div>