+++ src/views/teacher/teacherCreate.php
@@ -1,94 +1,194 @@
 <?php require_once "./src/views/header.php"; ?>
 
-<body>
-  <?php require_once "./src/views/navbar.php"; ?>
+<div class="column is-12">
 
-  <h2>Espace professeur - Création d'une grille</h2>
-
-  <form action="" method="POST">
-    <label for="newGridNum">Numéro de grille :</label>
-    <input type="text" name="newGridNum" required>
-
-    <label for="newGridName">Nom de l'étudiant :</label>
-    <input type="text" name="newGridName" required>
-
-    <label for="newGridFirstname">Prénom de l'étudiant :</label>
-    <input type="text" name="newGridFirstname" required>
-
-    <label for="newGridDiploma">Diplôme de l'étudiant :</label>
-    <select name="newGridDiploma" required>
-      <option hidden value="">Choisissez une option</option>
-      <option value=12>ES/S</option>
-      <option value=10>STMG</option>
-      <option value=8>PRO</option>
-      <option value=9>L</option>
-      <option value=5>Autre</option>
-    </select>
-
-    <label for="newGridWork">Travail de l'étudiant :</label>
-    <select name="newGridWork" required>
-      <option hidden value="">Choisissez une option</option>
-      <option value=1>Correct</option>
-      <option value=-1>Incorrect</option>
-    </select>
-
-    <label for="newGridAbs">Taux d'absence de l'étudiant :</label>
-    <select name="newGridAbs" required>
-      <option hidden value="">Choisissez une option</option>
-      <option value=0>Bas</option>
-      <option value=-2>Élevé</option>
-    </select>
-
-    <label for="newGridAtt">Attitude / Comportement de l'étudiant :</label>
-    <select name="newGridAtt" required>
-      <option hidden value="">Choisissez une option</option>
-      <option value=0>Correct</option>
-      <option value=-3>Incorrect</option>
-    </select>
-
-    <label for="newGridStudy">Études supérieures ? :</label>
-    <select name="newGridStudy" required>
-      <option hidden value="">Choisissez une option</option>
-      <option value=1>Oui</option>
-      <option value=0>Non</option>
-    </select>
-
-    <label for="newGridPpview">Avis du professeur principal :</label>
-    <select name="newGridPpview" required>
-      <option hidden value="">Choisissez une option</option>
-      <option value=2>B</option>
-      <option value=1>AB</option>
-      <option value=0>Insuffisant</option>
-      <option value=-2>Négatif</option>
-    </select>
-
-    <label for="newGridProview">Avis du proviseur :</label>
-    <select name="newGridProview" required>
-      <option hidden value="">Choisissez une option</option>
-      <option value=2>B</option>
-      <option value=1>AB</option>
-      <option value=0>Insuffisant</option>
-      <option value=-2>Négatif</option>
-    </select>
-
-    <label for="newGridCoverletter">Qualité lettre de motivation :</label>
-    <select name="newGridCoverletter" required>
-      <option hidden value="">Choisissez une option</option>
-      <option value=2>B</option>
-      <option value=1>AB</option>
-      <option value=0>Insuffisant</option>
-      <option value=-2>Négatif</option>
-    </select>
-
-    <label for="newGridComment">Remarques :</label>
-    <textarea name="newGridComment" rows=10 cols=65></textarea>
-
-    <button type="submit">Enregistrer</button>
-  </form>
-
+  <!-- Header -->
+  <h3 class="title has-text-black">[S11] Selection BTS - Espace Professeur - Création</h3>
+  <hr class="login-hr">
+  <p class="subtitle has-text-black">
+    Vous pouvez créer une grille en complétant l'ensemble des informations demandées.
+  </p>
   <?php if (isset($_GET["failed"])) : ?>
-    <p>Veuillez vérifier les informations rentrées.</p>
+    <p class="field is-size-5 has-text-red">Veuillez vérifier les informations rentrées.</p>
   <?php endif; ?>
-</body>
+  <!-- ./Header -->
 
-</html>
+  <!-- Main content -->
+  <div class="columns is-multiline">
+
+    <!-- Buttons -->
+    <div class="column is-one-quarter">
+      <div class="box">
+        <a class="field button is-block is-dark is-medium" href="?logout">
+          Se déconnecter
+          <i class="mdi mdi-logout" aria-hidden="true"></i>
+        </a>
+        <a class="field button is-block is-dark is-medium" href="?">
+          Accueil
+          <i class="mdi mdi-home" aria-hidden="true"></i>
+        </a>
+      </div>
+    </div>
+    <!-- ./Buttons -->
+
+    <!-- Form -->
+    <div class="column">
+      <form action="" method="POST" class="box">
+        <div class="field">
+          <label class="label">Numéro de grille</label>
+          <div class="control">
+            <input class="input is-medium" name="newGridNum" type="text" placeholder="#ID" required />
+          </div>
+        </div>
+
+        <div class="field">
+          <label class="label">Nom de l'étudiant</label>
+          <div class="control">
+            <input class="input is-medium" name="newGridName" type="text" placeholder="Nom" required />
+          </div>
+        </div>
+
+        <div class="field">
+          <label class="label">Prénom de l'étudiant</label>
+          <div class="control">
+            <input class="input is-medium" name="newGridFirstname" type="text" placeholder="Prénom" required />
+          </div>
+        </div>
+
+        <div class="field">
+          <label class="label">Diplome de l'étudiant</label>
+          <div class="control">
+            <div class="select is-medium is-fullwidth">
+              <select name="newGridDiploma" required>
+                <option hidden value=""></option>
+                <option value=12>ES/S</option>
+                <option value=10>STMG</option>
+                <option value=8>PRO</option>
+                <option value=9>L</option>
+                <option value=5>Autre</option>
+              </select>
+            </div>
+          </div>
+        </div>
+
+        <div class="field">
+          <label class="label">Travail de l'étudiant</label>
+          <div class="control">
+            <div class="select is-medium is-fullwidth">
+              <select name="newGridWork" required>
+                <option hidden value=""></option>
+                <option value=1>Correct</option>
+                <option value=-1>Incorrect</option>
+              </select>
+            </div>
+          </div>
+        </div>
+
+        <div class="field">
+          <label class="label">Taux d'absence de l'étudiant</label>
+          <div class="control">
+            <div class="select is-medium is-fullwidth">
+              <select name="newGridAbs" required>
+                <option hidden value=""></option>
+                <option value=0>Bas</option>
+                <option value=-2>Élevé</option>
+              </select>
+            </div>
+          </div>
+        </div>
+
+        <div class="field">
+          <label class="label">Attitude / Comportement de l'étudiant</label>
+          <div class="control">
+            <div class="select is-medium is-fullwidth">
+              <select name="newGridAtt" required>
+                <option hidden value=""></option>
+                <option value=0>Correct</option>
+                <option value=-3>Incorrect</option>
+              </select>
+            </div>
+          </div>
+        </div>
+
+        <div class="field">
+          <label class="label">Études supérieures ?</label>
+          <div class="control">
+            <div class="select is-medium is-fullwidth">
+              <select name="newGridStudy" required>
+                <option hidden value=""></option>
+                <option value=1>Oui</option>
+                <option value=0>Non</option>
+              </select>
+            </div>
+          </div>
+        </div>
+
+        <div class="field">
+          <label class="label">Avis du professeur principal</label>
+          <div class="control">
+            <div class="select is-medium is-fullwidth">
+              <select name="newGridPpview" required>
+                <option hidden value=""></option>
+                <option value=2>B</option>
+                <option value=1>AB</option>
+                <option value=0>Insuffisant</option>
+                <option value=-2>Négatif</option>
+              </select>
+            </div>
+          </div>
+        </div>
+
+        <div class="field">
+          <label class="label">Avis du proviseur</label>
+          <div class="control">
+            <div class="select is-medium is-fullwidth">
+              <select name="newGridProview" required>
+                <option hidden value=""></option>
+                <option value=2>B</option>
+                <option value=1>AB</option>
+                <option value=0>Insuffisant</option>
+                <option value=-2>Négatif</option>
+              </select>
+            </div>
+          </div>
+        </div>
+
+        <div class="field">
+          <label class="label">Qualité lettre de motivation</label>
+          <div class="control">
+            <div class="select is-medium is-fullwidth">
+              <select name="newGridCoverletter" required>
+                <option hidden value=""></option>
+                <option value=2>B</option>
+                <option value=1>AB</option>
+                <option value=0>Insuffisant</option>
+                <option value=-2>Négatif</option>
+              </select>
+            </div>
+          </div>
+        </div>
+
+        <div class="field">
+          <label class="label">Remarques</label>
+          <div class="control">
+            <textarea name="newGridComment" class="textarea is-medium is-fullwidth"></textarea>
+          </div>
+        </div>
+
+        <button type="submit" class="button is-block is-dark is-medium is-fullwidth">
+          Valider la création
+        </button>
+      </form>
+    </div>
+    <!-- ./Form -->
+
+  </div>
+  <!-- ./Main content -->
+
+  <!-- Footer -->
+  <p class="has-text-grey">NOIZET Maxence &nbsp;·&nbsp; 2021</p>
+  <!-- ./Footer -->
+
+</div>
+
+<?php require_once "./src/views/footer.php"; ?>